move generator to .mts extension
