+++ packages/generator/package.json
@@ -20,5 +20,6 @@
     "typescript": "^4.7.4",
     "uvu": "^0.5.6",
     "watchlist": "^0.3.1"
-  }
+  },
+  "type": "module"
 }
