+++ packages/generator/src/constants.mts
