+++ packages/generator/src/files/generate-adapter-angular.mts
