+++ packages/generator/src/files/generate-adapter-node.mts
