+++ packages/generator/src/files/generate-adapter-react.mts
