+++ packages/generator/src/files/generate-adapter-solid.mts
