+++ packages/generator/src/files/generate-adapter-svelte.mts
