+++ packages/generator/src/files/generate-adapter-vue.mts
