+++ packages/generator/src/files/generate-template-formatters.mts
