+++ packages/generator/src/files/generate-template-locale.mts
@@ -2,15 +2,15 @@ import type { BaseTranslation, Locale } from '@typesafe-i18n/runtime/core.mjs'
 import { join } from 'path'
 import type { GeneratorConfigWithDefaultValues } from '../../../config/src/types.mjs'
 import { fileEnding } from '../output-handler.mjs'
-import { getDictionaryTemplate } from '../utils/dictionary.utils'
+import { getDictionaryTemplate } from '../utils/dictionary.utils.mjs'
 import { writeFileIfContainsChanges } from '../utils/file.utils.mjs'
 import { prettify } from '../utils/generator.utils.mjs'
-import { logger } from '../utils/logger.mjs'
 
-export const generateNamespaceTemplate = async (
+// --------------------------------------------------------------------------------------------------------------------
+
+export const generateLocaleTemplate = async (
 	config: GeneratorConfigWithDefaultValues,
 	locale: Locale,
-	namespace: string,
 	translations: BaseTranslation | BaseTranslation[] | undefined = undefined,
 	editHint = '',
 	showBanner = false,
@@ -22,14 +22,12 @@ export const generateNamespaceTemplate = async (
 	const localeTemplate = getDictionaryTemplate(
 		config,
 		locale,
-		namespace,
+		undefined,
 		isBaseLocale,
 		translations,
 		editHint,
 		showBanner,
 	)
 
-	logger.info(`creating boilerplate for locale '${locale}' and namespace '${namespace}'`)
-
-	await writeFileIfContainsChanges(join(outputPath, locale, namespace), `index${fileEnding}`, prettify(localeTemplate))
+	await writeFileIfContainsChanges(join(outputPath, locale), `index${fileEnding}`, prettify(localeTemplate))
 }
