+++ packages/generator/src/files/generate-template-types.mts
