+++ packages/generator/src/files/generate-util-sync.mts
