+++ packages/generator/src/files/generate-util.mts
