+++ packages/generator/src/generate-files.mts
@@ -1,23 +1,23 @@
 import type { BaseTranslation, Locale } from '@typesafe-i18n/runtime/core.mjs'
 import type { GeneratorConfigWithDefaultValues } from '../../config/src/types.mjs'
-import { generateAngularAdapter } from './files/generate-adapter-angular'
-import { generateNodeAdapter } from './files/generate-adapter-node'
-import { generateReactAdapter } from './files/generate-adapter-react'
-import { generateSolidAdapter } from './files/generate-adapter-solid'
-import { generateSvelteAdapter } from './files/generate-adapter-svelte'
-import { generateVueAdapter } from './files/generate-adapter-vue'
-import { generateFormattersTemplate } from './files/generate-template-formatters'
-import { generateLocaleTemplate } from './files/generate-template-locale'
-import { generateNamespaceTemplate } from './files/generate-template-namespace'
-import { generateCustomTypesTemplate } from './files/generate-template-types'
-import { generateTypes } from './files/generate-types'
-import { generateUtil } from './files/generate-util'
-import { generateAsyncUtil } from './files/generate-util-async'
-import { generateSyncUtil } from './files/generate-util-sync'
+import { generateAngularAdapter } from './files/generate-adapter-angular.mjs'
+import { generateNodeAdapter } from './files/generate-adapter-node.mjs'
+import { generateReactAdapter } from './files/generate-adapter-react.mjs'
+import { generateSolidAdapter } from './files/generate-adapter-solid.mjs'
+import { generateSvelteAdapter } from './files/generate-adapter-svelte.mjs'
+import { generateVueAdapter } from './files/generate-adapter-vue.mjs'
+import { generateFormattersTemplate } from './files/generate-template-formatters.mjs'
+import { generateLocaleTemplate } from './files/generate-template-locale.mjs'
+import { generateNamespaceTemplate } from './files/generate-template-namespace.mjs'
+import { generateCustomTypesTemplate } from './files/generate-template-types.mjs'
+import { generateTypes } from './files/generate-types.mjs'
+import { generateAsyncUtil } from './files/generate-util-async.mjs'
+import { generateSyncUtil } from './files/generate-util-sync.mjs'
+import { generateUtil } from './files/generate-util.mjs'
 import { configureOutputHandler } from './output-handler.mjs'
 import type { TypescriptVersion } from './utils/generator.utils.mjs'
 import { logger as defaultLogger, Logger } from './utils/logger.mjs'
-import { findAllNamespacesForLocale } from './utils/namespaces.utils'
+import { findAllNamespacesForLocale } from './utils/namespaces.utils.mjs'
 
 export const generateNamespaceFiles = async (
 	config: GeneratorConfigWithDefaultValues,
