+++ packages/generator/src/parse-language-file.mts
