+++ packages/generator/src/utils/dictionary.utils.mts
@@ -8,10 +8,10 @@ import {
 	jsDocType,
 	relativeFileImportPath,
 	tsCheck,
-	type,
+	type
 } from '../output-handler.mjs'
 import { sanitizePath, wrapObjectKeyIfNeeded } from './generator.utils.mjs'
-import { getTypeNameForNamespace } from './namespaces.utils'
+import { getTypeNameForNamespace } from './namespaces.utils.mjs'
 
 // --------------------------------------------------------------------------------------------------------------------
