+++ packages/generator/src/utils/namespaces.utils.mts
