+++ packages/generator/test/wrap-strings.test.ts
@@ -1,6 +1,6 @@
 import { suite } from 'uvu'
 import * as assert from 'uvu/assert'
-import { getWrappedString } from '../src/utils/dictionary.utils'
+import { getWrappedString } from '../src/utils/dictionary.utils.mjs'
 
 const test = suite('wrap-strings')
