+++ packages/importer/src/importer.ts
@@ -1,11 +1,11 @@
 import type { BaseTranslation, ImportLocaleMapping, Locale } from '@typesafe-i18n/runtime/core.mjs'
 import ts from 'typescript'
-import { getConfigWithDefaultValues } from '../../config/src/config'
-import { generateLocaleTemplate } from '../../generator/src/files/generate-template-locale'
-import { generateNamespaceTemplate } from '../../generator/src/files/generate-template-namespace'
-import { generate } from '../../generator/src/generate-files'
+import { getConfigWithDefaultValues } from '../../config/src/config.mjs'
+import { generateLocaleTemplate } from '../../generator/src/files/generate-template-locale.mjs'
+import { generateNamespaceTemplate } from '../../generator/src/files/generate-template-namespace.mjs'
+import { generate } from '../../generator/src/generate-files.mjs'
 import { configureOutputHandler } from '../../generator/src/output-handler.mjs'
-import { getAllLocales, parseLanguageFile } from '../../generator/src/parse-language-file'
+import { getAllLocales, parseLanguageFile } from '../../generator/src/parse-language-file.mjs'
 import { parseTypescriptVersion } from '../../generator/src/utils/generator.utils.mjs'
 import { createLogger } from '../../generator/src/utils/logger.mjs'
